pyproject.toml
[project]
name = "gtmforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
